Add changes base algorithm

+++ back_end/internal/models/needs.go
@@ -6,7 +6,7 @@ type Need struct {
 	CounterpartyName  string  `json:"counterparty_name"`   // наименование контрагента
 	CollateralNorm    string  `json:"collateral_norm"`     // норма обеспечения
 	Property          string  `json:"property"`            // имущество
-	Need              float64 `json:"need"`                // потребность
+	Need              uint8   `json:"need"`                // потребность
 	WornOut           float64 `json:"worn_out"`            // изношено
 	NotWrittenOff     float64 `json:"not_written_off"`     // не списано
 	StartDate         string  `json:"start_date"`          // дата начала
